refactor(tests): restructure dirs, remove redundant tests
- Move pytest tests to tests/integration/
- Move golden rules to tests/golden_rules/
- Remove redundant test_mytool_help.py tests:
  TestMytoolFilesExist, TestMytoolFunctionality,
  TestMytoolIntegration (replaced by bats/duplicates)
- Remove duplicate mytool tests from compatibility
- Update runner, tox.ini, AGENTS.md, conftest path
