+++ tests/golden_rules/test_golden_rules.sh
