+++ tests/integration/test_file_cleanup.py
