+++ tox.ini
@@ -108,6 +108,13 @@ commands =
     shfmt -d -i 4 bash/    
 
 
+[testenv:bats]
+description = Run bats shell unit tests
+skip_install = true
+allowlist_externals = tests/bats/lib/bats-core/bin/bats
+commands = tests/bats/lib/bats-core/bin/bats tests/bats/init tests/bats/functions tests/bats/tools
+
+
 # --- 파이썬 버전별 테 스트 환경 ---
 # pyproject.toml의 requires-python 및 black.target-version과 일치시킵니다.
 # py310, py311, py312, py313은 각 파이썬 버전별 테스트 환경.
